--- examples/vision/prototypical_network_omniglot.py
@@ -1,737 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-"""
-Reproduce Omniglot results of Snell et al Prototypical networks.
-"""
-from torch.optim import Adam
-from torch.utils.data import DataLoader
-from torch.utils.data import Dataset
-import argparse
-from torch import nn
-from typing import List, Iterable, Callable, Tuple, Union
-from torch.utils.data import Sampler
-from tqdm import tqdm
-import os
-import torch
-from torch.optim import Optimizer
-from torch.nn import Module
-from typing import Callable
-from torchvision.datasets import Omniglot
-from torchvision import transforms
-from PIL.Image import LANCZOS
-from torch.utils.data import ConcatDataset
-import pandas as pd
-import numpy as np
-
-PATH = os.path.dirname(os.path.realpath(__file__))
-
-DATA_PATH = '/datadrive/few-shot/omniglot/python'
-
-EPSILON = 1e-8
-
-def mkdir(dir_name):
-    try:
-        os.mkdir(dir_name)
-    except:
-        pass
-
-def setup_dirs():
-    """Creates directories for this project."""
-    mkdir(PATH + '/logs/')
-    mkdir(PATH + '/logs/proto_nets')
-    mkdir(PATH + '/logs/matching_nets')
-    mkdir(PATH + '/logs/maml')
-    mkdir(PATH + '/models/')
-    mkdir(PATH + '/models/proto_nets')
-    mkdir(PATH + '/models/matching_nets')
-    mkdir(PATH + '/models/maml')
-
-
-    
-def categorical_accuracy(y, y_pred):
-    return torch.eq(y_pred.argmax(dim=-1), y).sum().item() / y_pred.shape[0]
-
-
-setup_dirs()
-assert torch.cuda.is_available()
-device = torch.device('cuda')
-torch.backends.cudnn.benchmark = True
-
-
-##############
-# Parameters #
-##############
-parser = argparse.ArgumentParser()
-parser.add_argument('--dataset')
-parser.add_argument('--distance', default='l2')
-parser.add_argument('--n-train', default=1, type=int)
-parser.add_argument('--n-test', default=1, type=int)
-parser.add_argument('--k-train', default=60, type=int)
-parser.add_argument('--k-test', default=5, type=int)
-parser.add_argument('--q-train', default=5, type=int)
-parser.add_argument('--q-test', default=1, type=int)
-args = parser.parse_args("".split())
-
-evaluation_episodes = 1000
-episodes_per_epoch = 100
-
-args.dataset = 'omniglot'
-
-class OmniglotDataset(Dataset):
-    def __init__(self, subset):
-        """Dataset class representing Omniglot dataset
-
-        # Arguments:
-            subset: Whether the dataset represents the background or evaluation set
-        """
-        if subset not in ('background', 'evaluation'):
-            raise(ValueError, 'subset must be one of (background, evaluation)')
-        self.subset = subset
-
-        self.df = pd.DataFrame(self.index_subset(self.subset))
-
-        # Index of dataframe has direct correspondence to item in dataset
-        self.df = self.df.assign(id=self.df.index.values)
-
-        # Convert arbitrary class names of dataset to ordered 0-(num_speakers - 1) integers
-        self.unique_characters = sorted(self.df['class_name'].unique())
-        self.class_name_to_id = {self.unique_characters[i]: i for i in range(self.num_classes())}
-        self.df = self.df.assign(class_id=self.df['class_name'].apply(lambda c: self.class_name_to_id[c]))
-
-        # Create dicts
-        self.datasetid_to_filepath = self.df.to_dict()['filepath']
-        self.datasetid_to_class_id = self.df.to_dict()['class_id']
-
-    def __getitem__(self, item):
-        instance = io.imread(self.datasetid_to_filepath[item])
-        # Reindex to channels first format as supported by pytorch
-        instance = instance[np.newaxis, :, :]
-
-        # Normalise to 0-1
-        instance = (instance - instance.min()) / (instance.max() - instance.min())
-
-        label = self.datasetid_to_class_id[item]
-
-        return torch.from_numpy(instance), label
-
-    def __len__(self):
-        return len(self.df)
-
-    def num_classes(self):
-        return len(self.df['class_name'].unique())
-
-    @staticmethod
-    def index_subset(subset):
-        """Index a subset by looping through all of its files and recording relevant information.
-
-        # Arguments
-            subset: Name of the subset
-
-        # Returns
-            A list of dicts containing information about all the image files in a particular subset of the
-            Omniglot dataset dataset
-        """
-        images = []
-        print('Indexing {}...'.format(subset))
-        # Quick first pass to find total for tqdm bar
-        subset_len = 0
-        for root, folders, files in os.walk(DATA_PATH + '/Omniglot/images_{}/'.format(subset)):
-            subset_len += len([f for f in files if f.endswith('.png')])
-
-        progress_bar = tqdm(total=subset_len)
-        for root, folders, files in os.walk(DATA_PATH + '/Omniglot/images_{}/'.format(subset)):
-            if len(files) == 0:
-                continue
-
-            alphabet = root.split('/')[-2]
-            class_name = '{}.{}'.format(alphabet, root.split('/')[-1])
-
-            for f in files:
-                progress_bar.update(1)
-                images.append({
-                    'subset': subset,
-                    'alphabet': alphabet,
-                    'class_name': class_name,
-                    'filepath': os.path.join(root, f)
-                })
-
-        progress_bar.close()
-        return images
-
-if args.dataset == 'omniglot':
-    n_epochs = 40
-    dataset_class = OmniglotDataset
-    num_input_channels = 1
-    drop_lr_every = 20
-elif args.dataset == 'miniImageNet':
-    n_epochs = 80
-    dataset_class = MiniImageNet
-    num_input_channels = 3
-    drop_lr_every = 40
-else:
-    raise(ValueError, 'Unsupported dataset')
-
-param_str = f'{args.dataset}_nt={args.n_train}_kt={args.k_train}_qt={args.q_train}_'             f'nv={args.n_test}_kv={args.k_test}_qv={args.q_test}'
-
-print(param_str)
-
-###################
-# Create datasets #
-###################
-
-class NShotTaskSampler(Sampler):
-    def __init__(self,
-                 dataset: torch.utils.data.Dataset,
-                 episodes_per_epoch: int = None,
-                 n: int = None,
-                 k: int = None,
-                 q: int = None,
-                 num_tasks: int = 1,
-                 fixed_tasks: List[Iterable[int]] = None):
-        """PyTorch Sampler subclass that generates batches of n-shot, k-way, q-query tasks.
-
-        Each n-shot task contains a "support set" of `k` sets of `n` samples and a "query set" of `k` sets
-        of `q` samples. The support set and the query set are all grouped into one Tensor such that the first n * k
-        samples are from the support set while the remaining q * k samples are from the query set.
-
-        The support and query sets are sampled such that they are disjoint i.e. do not contain overlapping samples.
-
-        # Arguments
-            dataset: Instance of torch.utils.data.Dataset from which to draw samples
-            episodes_per_epoch: Arbitrary number of batches of n-shot tasks to generate in one epoch
-            n_shot: int. Number of samples for each class in the n-shot classification tasks.
-            k_way: int. Number of classes in the n-shot classification tasks.
-            q_queries: int. Number query samples for each class in the n-shot classification tasks.
-            num_tasks: Number of n-shot tasks to group into a single batch
-            fixed_tasks: If this argument is specified this Sampler will always generate tasks from
-                the specified classes
-        """
-        super(NShotTaskSampler, self).__init__(dataset)
-        self.episodes_per_epoch = episodes_per_epoch
-        self.dataset = dataset
-        if num_tasks < 1:
-            raise ValueError('num_tasks must be > 1.')
-
-        self.num_tasks = num_tasks
-        # TODO: Raise errors if initialise badly
-        self.k = k
-        self.n = n
-        self.q = q
-        self.fixed_tasks = fixed_tasks
-
-        self.i_task = 0
-
-    def __len__(self):
-        return self.episodes_per_epoch
-
-    def __iter__(self):
-        for _ in range(self.episodes_per_epoch):
-            batch = []
-
-            for task in range(self.num_tasks):
-                if self.fixed_tasks is None:
-                    # Get random classes
-                    episode_classes = np.random.choice(self.dataset.df['class_id'].unique(), size=self.k, replace=False)
-                else:
-                    # Loop through classes in fixed_tasks
-                    episode_classes = self.fixed_tasks[self.i_task % len(self.fixed_tasks)]
-                    self.i_task += 1
-
-                df = self.dataset.df[self.dataset.df['class_id'].isin(episode_classes)]
-
-                support_k = {k: None for k in episode_classes}
-                for k in episode_classes:
-                    # Select support examples
-                    support = df[df['class_id'] == k].sample(self.n)
-                    support_k[k] = support
-
-                    for i, s in support.iterrows():
-                        batch.append(s['id'])
-
-                for k in episode_classes:
-                    query = df[(df['class_id'] == k) & (~df['id'].isin(support_k[k]['id']))].sample(self.q)
-                    for i, q in query.iterrows():
-                        batch.append(q['id'])
-
-            yield np.stack(batch)
-
-
-omni_background = Omniglot(root='./data',
-                               background=True,
-                               transform=transforms.Compose([
-                                   transforms.Resize(28, interpolation=LANCZOS),
-                                   transforms.ToTensor(),
-                                   # TODO: Add DiscreteRotations([0, 90, 180, 270])
-                                   lambda x: 1.0 - x,
-                               ]),
-                               download=True)
-
-omni_evaluation = Omniglot(root='./data',
-                               background=False,
-                               transform=transforms.Compose([
-                                   transforms.Resize(28, interpolation=LANCZOS),
-                                   transforms.ToTensor(),
-                                   # TODO: Add DiscreteRotations([0, 90, 180, 270])
-                                   lambda x: 1.0 - x,
-                               ]),
-                               target_transform=transforms.Compose([
-                                   lambda x: max_y + x,
-                               ]),
-                               download=True)
-omniglot = ConcatDataset((omni_background, omni_evaluation))
-
-from training_helper import *
-from skimage import io
-background = dataset_class('background')
-background_taskloader = DataLoader(
-    background,
-    batch_sampler=NShotTaskSampler(background, episodes_per_epoch, args.n_train, args.k_train, args.q_train),
-    num_workers=4
-)
-evaluation = dataset_class('evaluation')
-evaluation_taskloader = DataLoader(
-    evaluation,
-    batch_sampler=NShotTaskSampler(evaluation, episodes_per_epoch, args.n_test, args.k_test, args.q_test),
-    num_workers=4
-)
-
-
-
-
-
-def gradient_step(model: Module, optimiser: Optimizer, loss_fn: Callable, x: torch.Tensor, y: torch.Tensor, **kwargs):
-    model.train()
-    optimiser.zero_grad()
-    y_pred = model(x)
-    loss = loss_fn(y_pred, y)
-    loss.backward()
-    optimiser.step()
-
-    return loss, y_pred
-
-
-NAMED_METRICS = {
-    'categorical_accuracy': categorical_accuracy
-}
-
-
-
-def conv_block(in_channels: int, out_channels: int) -> nn.Module:
-    return nn.Sequential(
-        nn.Conv2d(in_channels, out_channels, 3, padding=1),
-        nn.BatchNorm2d(out_channels),
-        nn.ReLU(),
-        nn.MaxPool2d(kernel_size=2, stride=2)
-    )
-
-class Flatten(nn.Module):
-    def forward(self, input):
-        return input.view(input.size(0), -1)
-    
-def lr_schedule(epoch, lr):
-    # Drop lr every 2000 episodes
-    if epoch % drop_lr_every == 0:
-        return lr / 2
-    else:
-        return lr
-
-
-model = nn.Sequential(
-        conv_block(num_input_channels, 64),
-        conv_block(64, 64),
-        conv_block(64, 64),
-        conv_block(64, 64),
-        Flatten(),
-    )
-model.to(device, dtype=torch.double)
-
-
-optimiser = Adam(model.parameters(), lr=1e-3)
-loss_fn = torch.nn.NLLLoss().cuda()
-
-
-
-class Callback(object):
-    def __init__(self):
-        self.model = None
-
-    def set_params(self, params):
-        self.params = params
-
-    def set_model(self, model):
-        self.model = model
-
-    def on_epoch_begin(self, epoch, logs=None):
-        pass
-
-    def on_epoch_end(self, epoch, logs=None):
-        pass
-
-    def on_batch_begin(self, batch, logs=None):
-        pass
-
-    def on_batch_end(self, batch, logs=None):
-        pass
-
-    def on_train_begin(self, logs=None):
-        pass
-
-    def on_train_end(self, logs=None):
-        pass
-
-
-class EvaluateFewShot(Callback):
-    """Evaluate a network on  an n-shot, k-way classification tasks after every epoch.
-
-    # Arguments
-        eval_fn: Callable to perform few-shot classification. Examples include `proto_net_episode`,
-            `matching_net_episode` and `meta_gradient_step` (MAML).
-        num_tasks: int. Number of n-shot classification tasks to evaluate the model with.
-        n_shot: int. Number of samples for each class in the n-shot classification tasks.
-        k_way: int. Number of classes in the n-shot classification tasks.
-        q_queries: int. Number query samples for each class in the n-shot classification tasks.
-        task_loader: Instance of NShotWrapper class
-        prepare_batch: function. The preprocessing function to apply to samples from the dataset.
-        prefix: str. Prefix to identify dataset.
-    """
-
-    def __init__(self,
-                 eval_fn: Callable,
-                 num_tasks: int,
-                 n_shot: int,
-                 k_way: int,
-                 q_queries: int,
-                 taskloader: torch.utils.data.DataLoader,
-                 prepare_batch: Callable,
-                 prefix: str = 'val_',
-                 **kwargs):
-        super(EvaluateFewShot, self).__init__()
-        self.eval_fn = eval_fn
-        self.num_tasks = num_tasks
-        self.n_shot = n_shot
-        self.k_way = k_way
-        self.q_queries = q_queries
-        self.taskloader = taskloader
-        self.prepare_batch = prepare_batch
-        self.prefix = prefix
-        self.kwargs = kwargs
-        self.metric_name = f'{self.prefix}{self.n_shot}-shot_{self.k_way}-way_acc'
-
-    def on_train_begin(self, logs=None):
-        self.loss_fn = self.params['loss_fn']
-        self.optimiser = self.params['optimiser']
-
-    def on_epoch_end(self, epoch, logs=None):
-        logs = logs or {}
-        seen = 0
-        totals = {'loss': 0, self.metric_name: 0}
-        for batch_index, batch in enumerate(self.taskloader):
-            x, y = self.prepare_batch(batch)
-
-            loss, y_pred = self.eval_fn(
-                self.model,
-                self.optimiser,
-                self.loss_fn,
-                x,
-                y,
-                n_shot=self.n_shot,
-                k_way=self.k_way,
-                q_queries=self.q_queries,
-                train=False,
-                **self.kwargs
-            )
-
-            seen += y_pred.shape[0]
-
-            totals['loss'] += loss.item() * y_pred.shape[0]
-            totals[self.metric_name] += categorical_accuracy(y, y_pred) * y_pred.shape[0]
-
-        logs[self.prefix + 'loss'] = totals['loss'] / seen
-        logs[self.metric_name] = totals[self.metric_name] / seen
-
-
-        
-
-def fit(model: Module, optimiser: Optimizer, loss_fn: Callable, epochs: int, dataloader: DataLoader,
-        prepare_batch: Callable, metrics: List[Union[str, Callable]] = None, callbacks: List[Callback] = None,
-        verbose: bool =True, fit_function: Callable = gradient_step, fit_function_kwargs: dict = {}):
-
-    # Determine number of samples:
-    num_batches = len(dataloader)
-    batch_size = dataloader.batch_size
-
-    callbacks = CallbackList([DefaultCallback(), ] + (callbacks or []) + [ProgressBarLogger(), ])
-    callbacks.set_model(model)
-    callbacks.set_params({
-        'num_batches': num_batches,
-        'batch_size': batch_size,
-        'verbose': verbose,
-        'metrics': (metrics or []),
-        'prepare_batch': prepare_batch,
-        'loss_fn': loss_fn,
-        'optimiser': optimiser
-    })
-
-    if verbose:
-        print('Begin training...')
-
-    callbacks.on_train_begin()
-
-    for epoch in range(1, epochs+1):
-        callbacks.on_epoch_begin(epoch)
-
-        epoch_logs = {}
-        for batch_index, batch in enumerate(dataloader):
-            batch_logs = dict(batch=batch_index, size=(batch_size or 1))
-
-            callbacks.on_batch_begin(batch_index, batch_logs)
-
-            x, y = prepare_batch(batch)
-
-            loss, y_pred = fit_function(model, optimiser, loss_fn, x, y, **fit_function_kwargs)
-            batch_logs['loss'] = loss.item()
-
-            # Loops through all metrics
-            batch_logs = batch_metrics(model, y_pred, y, metrics, batch_logs)
-
-            callbacks.on_batch_end(batch_index, batch_logs)
-
-        # Run on epoch end
-        callbacks.on_epoch_end(epoch, epoch_logs)
-
-    # Run on train end
-    if verbose:
-        print('Finished.')
-
-    callbacks.on_train_end()
-
-
-        
-def prepare_nshot_task(n: int, k: int, q: int) -> Callable:
-    """Typical n-shot task preprocessing.
-
-    # Arguments
-        n: Number of samples for each class in the n-shot classification task
-        k: Number of classes in the n-shot classification task
-        q: Number of query samples for each class in the n-shot classification task
-
-    # Returns
-        prepare_nshot_task_: A Callable that processes a few shot tasks with specified n, k and q
-    """
-    def prepare_nshot_task_(batch: Tuple[torch.Tensor, torch.Tensor]) -> Tuple[torch.Tensor, torch.Tensor]:
-        """Create 0-k label and move to GPU.
-
-        TODO: Move to arbitrary device
-        """
-        x, y = batch
-        x = x.double().cuda()
-        # Create dummy 0-(num_classes - 1) label
-        y = create_nshot_task_label(k, q).cuda()
-        return x, y
-
-    return prepare_nshot_task_
-
-
-def batch_metrics(model: Module, y_pred: torch.Tensor, y: torch.Tensor, metrics: List[Union[str, Callable]],
-                  batch_logs: dict):
-    """Calculates metrics for the current training batch
-
-    # Arguments
-        model: Model being fit
-        y_pred: predictions for a particular batch
-        y: labels for a particular batch
-        batch_logs: Dictionary of logs for the current batch
-    """
-    model.eval()
-    for m in metrics:
-        if isinstance(m, str):
-            batch_logs[m] = NAMED_METRICS[m](y, y_pred)
-        else:
-            # Assume metric is a callable function
-            batch_logs = m(y, y_pred)
-
-    return batch_logs
-
-
-
-
-def proto_net_episode(model: Module,
-                      optimiser: Optimizer,
-                      loss_fn: Callable,
-                      x: torch.Tensor,
-                      y: torch.Tensor,
-                      n_shot: int,
-                      k_way: int,
-                      q_queries: int,
-                      distance: str,
-                      train: bool):
-    """Performs a single training episode for a Prototypical Network.
-
-    # Arguments
-        model: Prototypical Network to be trained.
-        optimiser: Optimiser to calculate gradient step
-        loss_fn: Loss function to calculate between predictions and outputs. Should be cross-entropy
-        x: Input samples of few shot classification task
-        y: Input labels of few shot classification task
-        n_shot: Number of examples per class in the support set
-        k_way: Number of classes in the few shot classification task
-        q_queries: Number of examples per class in the query set
-        distance: Distance metric to use when calculating distance between class prototypes and queries
-        train: Whether (True) or not (False) to perform a parameter update
-
-    # Returns
-        loss: Loss of the Prototypical Network on this task
-        y_pred: Predicted class probabilities for the query set on this task
-    """
-    if train:
-        # Zero gradients
-        model.train()
-        optimiser.zero_grad()
-    else:
-        model.eval()
-
-    # Embed all samples
-    embeddings = model(x)
-
-    # Samples are ordered by the NShotWrapper class as follows:
-    # k lots of n support samples from a particular class
-    # k lots of q query samples from those classes
-    support = embeddings[:n_shot*k_way]
-    queries = embeddings[n_shot*k_way:]
-    prototypes = compute_prototypes(support, k_way, n_shot)
-
-    # Calculate squared distances between all queries and all prototypes
-    # Output should have shape (q_queries * k_way, k_way) = (num_queries, k_way)
-    distances = pairwise_distances(queries, prototypes, distance)
-
-    # Calculate log p_{phi} (y = k | x)
-    log_p_y = (-distances).log_softmax(dim=1)
-    loss = loss_fn(log_p_y, y)
-
-    # Prediction probabilities are softmax over distances
-    y_pred = (-distances).softmax(dim=1)
-
-    if train:
-        # Take gradient step
-        loss.backward()
-        optimiser.step()
-    else:
-        pass
-
-    return loss, y_pred
-
-
-def compute_prototypes(support: torch.Tensor, k: int, n: int) -> torch.Tensor:
-    """Compute class prototypes from support samples.
-
-    # Arguments
-        support: torch.Tensor. Tensor of shape (n * k, d) where d is the embedding
-            dimension.
-        k: int. "k-way" i.e. number of classes in the classification task
-        n: int. "n-shot" of the classification task
-
-    # Returns
-        class_prototypes: Prototypes aka mean embeddings for each class
-    """
-    # Reshape so the first dimension indexes by class then take the mean
-    # along that dimension to generate the "prototypes" for each class
-    class_prototypes = support.reshape(k, n, -1).mean(dim=1)
-    return class_prototypes
-
-
-def create_nshot_task_label(k: int, q: int) -> torch.Tensor:
-    """Creates an n-shot task label.
-
-    Label has the structure:
-        [0]*q + [1]*q + ... + [k-1]*q
-
-    # TODO: Test this
-
-    # Arguments
-        k: Number of classes in the n-shot classification task
-        q: Number of query samples for each class in the n-shot classification task
-
-    # Returns
-        y: Label vector for n-shot task of shape [q * k, ]
-    """
-    y = torch.arange(0, k, 1 / q).long()
-    return y
-
-
-def pairwise_distances(x: torch.Tensor,
-                       y: torch.Tensor,
-                       matching_fn: str) -> torch.Tensor:
-    """Efficiently calculate pairwise distances (or other similarity scores) between
-    two sets of samples.
-
-    # Arguments
-        x: Query samples. A tensor of shape (n_x, d) where d is the embedding dimension
-        y: Class prototypes. A tensor of shape (n_y, d) where d is the embedding dimension
-        matching_fn: Distance metric/similarity score to compute between samples
-    """
-    n_x = x.shape[0]
-    n_y = y.shape[0]
-
-    if matching_fn == 'l2':
-        distances = (
-                x.unsqueeze(1).expand(n_x, n_y, -1) -
-                y.unsqueeze(0).expand(n_x, n_y, -1)
-        ).pow(2).sum(dim=2)
-        return distances
-    elif matching_fn == 'cosine':
-        normalised_x = x / (x.pow(2).sum(dim=1, keepdim=True).sqrt() + EPSILON)
-        normalised_y = y / (y.pow(2).sum(dim=1, keepdim=True).sqrt() + EPSILON)
-
-        expanded_x = normalised_x.unsqueeze(1).expand(n_x, n_y, -1)
-        expanded_y = normalised_y.unsqueeze(0).expand(n_x, n_y, -1)
-
-        cosine_similarities = (expanded_x * expanded_y).sum(dim=2)
-        return 1 - cosine_similarities
-    elif matching_fn == 'dot':
-        expanded_x = x.unsqueeze(1).expand(n_x, n_y, -1)
-        expanded_y = y.unsqueeze(0).expand(n_x, n_y, -1)
-
-        return -(expanded_x * expanded_y).sum(dim=2)
-    else:
-        raise(ValueError('Unsupported similarity function'))
-
-
-callbacks = [
-    EvaluateFewShot(
-        eval_fn=proto_net_episode,
-        num_tasks=evaluation_episodes,
-        n_shot=args.n_test,
-        k_way=args.k_test,
-        q_queries=args.q_test,
-        taskloader=evaluation_taskloader,
-        prepare_batch=prepare_nshot_task(args.n_test, args.k_test, args.q_test),
-        distance=args.distance
-    ),
-    ModelCheckpoint(
-        filepath=PATH + f'/models/proto_nets/{param_str}.pth',
-        monitor=f'val_{args.n_test}-shot_{args.k_test}-way_acc'
-    ),
-    LearningRateScheduler(schedule=lr_schedule),
-    CSVLogger(PATH + f'/logs/proto_nets/{param_str}.csv'),
-]
-
-if __name__ == '__main__':
-    fit(
-        model,
-        optimiser,
-        loss_fn,
-        epochs=n_epochs,
-        dataloader=background_taskloader,
-        prepare_batch=prepare_nshot_task(args.n_train, args.k_train, args.q_train),
-        callbacks=callbacks,
-        metrics=['categorical_accuracy'],
-        fit_function=proto_net_episode,
-        fit_function_kwargs={'n_shot': args.n_train, 'k_way': args.k_train, 'q_queries': args.q_train, 'train': True,
-                            'distance': args.distance},
-    )
-
-
-
-
-
